+++ lib/duckdb.rb
@@ -1,7 +1,7 @@
+require 'duckdb/duckdb_native'
 require 'duckdb/version'
 require 'duckdb/database'
 require 'duckdb/connection'
-require 'duckdb/duckdb_native'
 require 'duckdb/result'
 require 'duckdb/prepared_statement'