+++ ext/duckdb/prepared_statement.h
@@ -3,7 +3,7 @@
 
 struct _rubyDuckDBPreparedStatement {
     duckdb_prepared_statement prepared_statement;
-    index_t nparams;
+    idx_t nparams;
 };
 
 typedef struct _rubyDuckDBPreparedStatement rubyDuckDBPreparedStatement;